requirements.txt
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
PyYAML==6.0.2
wxPython==4.2.3
